+++ tests/test_function_registry.py
@@ -71,7 +71,7 @@ FUNCTIONS_INFO = [
 
 
 
-    ("ProjectP.py", "custom_helper_function", 27),
+    ("ProjectP.py", "custom_helper_function", 36),
 ]